notebook/llm-rag/code/requirements.txt
haystack-ai
sentence-transformers
accelerate
trafilatura
qdrant-haystack
fastembed-haystack
nltk
rerankers
prompt_toolkit